refactor: move GoogleMerchant admin logic to controllers
